norme - manque que les fonctions de plus de 25 lignes :))

+++ minishell-new/source/utils/utils_str.c
@@ -6,7 +6,7 @@
 /*   By: mprigent <mprigent@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/16 14:30:36 by sdesseau          #+#    #+#             */
-/*   Updated: 2022/04/21 22:20:19 by mprigent         ###   ########.fr       */
+/*   Updated: 2022/04/26 15:58:16 by mprigent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -83,14 +83,14 @@ size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 
 	i = 0;
 	j = 0;
-	if (dstsize < ft_strlen(dst))
-		return (dstsize + ft_strlen(src));
+	if (dstsize < (size_t)ft_strlen(dst))
+		return (dstsize + (size_t)ft_strlen(src));
 	if (dstsize == 0)
-		return (ft_strlen(src));
+		return ((size_t)ft_strlen(src));
 	while (dst[i] != '\0')
 		i++;
 	if (dstsize <= i)
-		return (dstsize + ft_strlen(src));
+		return (dstsize + (size_t)ft_strlen(src));
 	k = i;
 	while (src[j] && i < dstsize - 1)
 	{
@@ -99,7 +99,7 @@ size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 		j++;
 	}
 	dst[i] = '\0';
-	return (k + ft_strlen(src));
+	return (k + (size_t)ft_strlen(src));
 }
 
 char	*ft_strdup(const char *s1)
